# LUT optimizations