… plugins/obs-filters/data/LUTs/original.cube
… plugins/obs-filters/data/locale/en-US.ini