### Description
Two optimizations:
- Filter in linear space which is arguably more correct, and simplifies the shader.
- Replace tetrahedral interpolation with trilinear. Just use a bigger LUT.

This has a breaking change in that the Amount multiplier now applies in linear space, but I suspect most users use 1.0, so I don't think it's worth versioning the filter.

### Motivation and Context
Speed.

### How Has This Been Tested?
Linear filter, Intel UHD Graphics 750, 1080p draw: 980 µs -> 860 µs
Trilinear interpolation, Intel UHD Graphics 750, 1080p draw: 860 µs -> 591 µs
Remove domain checks, Intel UHD Graphics 750, 1080p draw: 591 µs -> 532 µs
Alpha passthrough, Intel UHD Graphics 750, 1080p draw: 532 µs -> 486 µs

Tested on D3D11 and OpenGL, and checked image diffs to make sure there weren't major regressions.

### Types of changes
- Performance enhancement (non-breaking change which improves efficiency)
- Breaking change (fix or feature that would cause existing functionality to change)

### Checklist:
- [x] My code has been run through [clang-format](https://github.com/obsproject/obs-studio/blob/master/.clang-format).
- [x] I have read the [**contributing** document](https://github.com/obsproject/obs-studio/blob/master/CONTRIBUTING.rst).
- [x] My code is not on the master branch.
- [x] The code has been tested.
- [x] All commit messages are properly formatted and commits squashed where appropriate.
- [x] I have included updates to all appropriate documentation.