@Nofew 2022-02-17
I've seen at least two people complain about the inverse of this change being done a while back, and both of them requested a toggle to change the behavior. This feels like a wonderful time to mention this, and suggest that such a toggle be added.