@Nofew 2022-02-17
I imagine if some people want the old behavior, there may also be some people who want the new behavior as well.