@jpark37 2022-02-17
Nah, I haven't heard anyone care either way. I just switched it over on a whim when adding CUBE support because the spec wanted it, but now I just want it to be fast.