@Vainock 2022-02-20
Could you describe in more detail what "Passthrough Alpha" does? I would like to provide more context on Crowdin as this is quite technical.