@jpark37 2022-02-20
The issue is that we're working with premultiplied colors, so a LUT conversion of the RGB channels can result in an invalid color. The RGB values should not exceed the alpha value. We can avoid this by dividing by alpha, performing the RGB LUT, and then remultiplying the alpha.

This checkbox allows the user to skip these extra calculations, e.g. because they know the incoming colors are opaque, which provides a modest speedup.