@SuslikV 2022-06-11
I think, the "Passthrough Alpha" setting should be hidden from the user.
Speed-up optimizations shouldn't depend on user (if they don't affect the quality).
The option shouldn't make incorrect results.

Notes:
1) I don't know if the check for opaque colors will preserve any benefits of the "Passthrough Alpha" optimization.
2) It is possible to transform this setting to "Make the source opaque", if it has any sense here.